gateway: split the truncation regressions for the test-module line budget
